+++ src/wininput.c
@@ -4,6 +4,7 @@
 
 #include "winpriv.h"
 #include "winsearch.h"
+#include "wintab.h"
 
 #include "charset.h"
 #include "child.h"
@@ -13,6 +14,7 @@
 #include <windowsx.h>  // GET_X_LPARAM, GET_Y_LPARAM
 #include <winnls.h>
 #include <termios.h>
+
 
 static HMENU ctxmenu = NULL;
 static HMENU sysmenu;
@@ -564,6 +566,14 @@ win_update_menus(bool callback)
     alt_fn ? W("Alt+F12") : ct_sh ? W("Ctrl+Shift+S") : null
   );
 
+  uint status_line = term.st_type == 1 ? MF_CHECKED
+                   : term.st_type == 0 ? MF_UNCHECKED
+                   : MF_GRAYED;
+  //__ Context menu:
+  modify_menu(ctxmenu, IDM_STATUSLINE, status_line, _W("Status Line"),
+    null
+  );
+
   uint options_enabled = config_wnd ? MF_GRAYED : MF_ENABLED;
   EnableMenuItem(ctxmenu, IDM_OPTIONS, options_enabled);
   EnableMenuItem(sysmenu, IDM_OPTIONS, options_enabled);
@@ -708,6 +718,7 @@ win_init_ctxmenu(bool extended_menu, bool with_user_commands)
   AppendMenuW(ctxmenu, MF_ENABLED | MF_UNCHECKED, IDM_SCROLLBAR, 0);
   AppendMenuW(ctxmenu, MF_ENABLED | MF_UNCHECKED, IDM_FULLSCREEN_ZOOM, 0);
   AppendMenuW(ctxmenu, MF_ENABLED | MF_UNCHECKED, IDM_FLIPSCREEN, 0);
+  AppendMenuW(ctxmenu, MF_ENABLED | MF_UNCHECKED, IDM_STATUSLINE, 0);
   AppendMenuW(ctxmenu, MF_SEPARATOR, 0, 0);
   if (extended_menu) {
     //__ Context menu: generate a TTY BRK condition (tty line interrupt)
@@ -1346,7 +1357,7 @@ toggle_status_line()
 {
   if (term.st_type == 1)
     term_set_status_type(0, 0);
-  else
+  else if (term.st_type == 0)
     term_set_status_type(1, 0);
 }
 
@@ -1671,7 +1682,9 @@ mflags_dim_margins()
 static uint
 mflags_status_line()
 {
-  return term.st_type == 1 ? MF_CHECKED : MF_UNCHECKED;
+  return term.st_type == 1 ? MF_CHECKED
+       : term.st_type == 0 ? MF_UNCHECKED
+       : MF_GRAYED;
 }
 
 static uint
@@ -1684,6 +1697,22 @@ static uint
 mflags_tek_mode()
 {
   return tek_mode ? MF_ENABLED : MF_GRAYED;
+}
+
+static uint
+mflags_tabbar()
+{
+  return win_tabbar_visible() ? MF_CHECKED : MF_UNCHECKED;
+}
+
+static void
+toggle_tabbar()
+{
+  cfg.tabbar = !cfg.tabbar;
+  if (cfg.tabbar)
+    win_open_tabbar();
+  else
+    win_close_tabbar();
 }
 
 static void hor_left_1() { horscroll(-1); }
@@ -1716,6 +1745,7 @@ static struct function_def cmd_defs[] = {
   {"new-window-cwd", {IDM_NEW_CWD}, 0},
   {"new-tab", {IDM_TAB}, 0},
   {"new-tab-cwd", {IDM_TAB_CWD}, 0},
+  {"toggle-tabbar", {.fct = toggle_tabbar}, mflags_tabbar},
 
   {"hor-left-1", {.fct = hor_left_1}, 0},
   {"hor-right-1", {.fct = hor_right_1}, 0},