+++ wiki/Changelog.md
@@ -1,3 +1,25 @@
+### 3.6.4 (25 Mar 2023) ###
+
+Terminal features
+  * Tweak response code to XTQALLOWED OSC 60 and XTQDISALLOWED OSC 61 (xterm 378).
+  * Readline mouse modes (DECSET 2001 / 2002 / 2003, xterm 379).
+  * Hover and open URL: support parentheses (#1196).
+  * Fix cursor artefacts in connection with ClicksPlaceCursor=yes.
+  * Fix mouse-paste while still selecting.
+  * Support multi-line progress detection.
+
+Windows integration
+  * Manage user-set HOME for calling Windows from WSL (mintty/wsltty#324, ~mintty/wsltty#76).
+  * Support OSC 7 directory cloning if cloning WSL window while in rootfs.
+
+Configuration
+  * Status line is configurable in Options menu, switchable from context menu.
+  * New user-definable function toggle-tabbar (#1201).
+  * Setting ClicksPlaceCursor presets all readline mouse modes.
+  * Dynamic support for flags emojis, deployment script (~mintty/wsltty#321).
+  * `MINTTY_DEBUG=C mintty ...` will list loaded config files (#1181).
+  * New option ProgressScan.
+
 ### 3.6.3 (18 Dec 2022) ###
 
 Terminal features
@@ -234,7 +256,7 @@ Configuration
   * New option BlinkColour.
   * New options MousePointer, AppMousePointer.
   * Restored "Allow blinking" in Options dialog (#1097).
-  * WSL-specific detection of Term info availabilities (mintty/wsltty#278).
+  * WSL-specific detection of available settings for Terminal Type (option Term) (mintty/wsltty#278).
   * Export TERM to WSL (mintty/wsltty#278).
 
 ### 3.4.7 (16 March 2021) ###